feat(trade): ✨ add composables for market status, trade execution, quote, tokens, and validation
- Implemented `useMarketStatus` for fetching and managing market status with countdown functionality.
- Created `useTradeExecution` for handling trade approvals and order submissions.
- Developed `useTradeQuote` to fetch trade quotes and manage approval requirements.
- Added `useTradeTokens` for managing tradable tokens and their selection logic.
- Introduced `useTradeValidation` to validate trade inputs and ensure compliance with trading rules.
