+++ src/modules/trade/composables/useTradeValidation.ts
@@ -1,0 +1,165 @@
+import { computed, type Ref, type ComputedRef } from 'vue'
+import { parseUnits } from 'viem'
+import BigNumber from 'bignumber.js'
+import type { NewTokenInfo } from '@/composables/useSwap'
+import { MAIN_TOKEN_CONTRACT, useWalletStore } from '@/stores/walletStore'
+import { useI18n } from 'vue-i18n'
+
+interface UseTradeValidationOptions {
+  fromTokenSelected: Ref<NewTokenInfo | null>
+  fromAmount: Ref<string>
+  toAmount: Ref<string>
+  isWalletConnected: Ref<boolean>
+  isMarketOpen: ComputedRef<boolean>
+  isSelectedAssetTradeable: ComputedRef<boolean>
+  supportedNetwork: ComputedRef<boolean>
+  isLoadingQuote: Ref<boolean>
+}
+
+export function useTradeValidation(options: UseTradeValidationOptions) {
+  const {
+    fromTokenSelected,
+    fromAmount,
+    toAmount,
+    isWalletConnected,
+    isMarketOpen,
+    isSelectedAssetTradeable,
+    supportedNetwork,
+    isLoadingQuote,
+  } = options
+
+  const { t } = useI18n()
+  const walletStore = useWalletStore()
+
+  // Helper to get token balance parameters
+  const getTokenBalanceParams = (token: NewTokenInfo) => {
+    const isMainToken = token.address === MAIN_TOKEN_CONTRACT
+    const balance = token.balance || '0'
+
+    const baseNetworkBalance = parseUnits(
+      walletStore.getTokenBalance(MAIN_TOKEN_CONTRACT)?.balance || '0',
+      18,
+    )
+
+    const decimals = token.decimals || 18
+    const baseBalance = isMainToken ? baseNetworkBalance : BigInt(balance)
+
+    return { baseBalance, decimals }
+  }
+
+  // Check if amount has pre-quote errors (excludes balance check)
+  const hasPreQuoteError = computed(() => {
+    if (
+      !fromTokenSelected.value ||
+      !fromAmount.value ||
+      fromAmount.value === '0'
+    ) {
+      return true
+    }
+
+    const amountBN = BigNumber(fromAmount.value)
+    if (amountBN.isNaN() || amountBN.lte(0)) {
+      return true
+    }
+
+    // Check decimals
+    const decimals = fromTokenSelected.value.decimals || 18
+    const decimalPlaces = fromAmount.value.includes('.')
+      ? fromAmount.value.split('.')[1]?.length || 0
+      : 0
+    if (decimalPlaces > decimals) {
+      return true
+    }
+
+    // Minimum $1 value check
+    const tokenPrice = fromTokenSelected.value.price || 0
+    if (tokenPrice > 0) {
+      const usdValue = amountBN
+        .times(tokenPrice)
+        .integerValue(BigNumber.ROUND_CEIL)
+      if (usdValue.lt(1)) {
+        return true
+      }
+    }
+
+    return false
+  })
+
+  // Computed error for from amount with balance validation
+  const fromAmountError = computed(() => {
+    if (
+      !fromTokenSelected.value ||
+      !fromAmount.value ||
+      fromAmount.value === '0'
+    ) {
+      return ''
+    }
+
+    const amountBN = BigNumber(fromAmount.value)
+    if (amountBN.isNaN()) {
+      return t('swap.error.invalid-amount')
+    }
+
+    if (amountBN.lte(0)) {
+      return t('swap.error.more-than-zero')
+    }
+
+    const decimals = fromTokenSelected.value.decimals || 18
+    const decimalPlaces = fromAmount.value.includes('.')
+      ? fromAmount.value.split('.')[1]?.length || 0
+      : 0
+    if (decimalPlaces > decimals) {
+      return t('swap.error.too-many-decimals')
+    }
+
+    // Minimum $1 value check
+    const tokenPrice = fromTokenSelected.value.price || 0
+    if (tokenPrice > 0) {
+      const usdValue = amountBN
+        .times(tokenPrice)
+        .integerValue(BigNumber.ROUND_CEIL)
+      if (usdValue.lt(1)) {
+        return 'Minimum trade value is $1.00'
+      }
+    }
+
+    // Balance check (only when wallet is connected)
+    if (isWalletConnected.value) {
+      try {
+        const tokenParams = getTokenBalanceParams(fromTokenSelected.value)
+        const baseAmount = parseUnits(amountBN.toFixed(decimals), decimals)
+
+        if (tokenParams.baseBalance < baseAmount) {
+          return t('swap.error.insufficient-native', {
+            symbol: fromTokenSelected.value.symbol,
+          })
+        }
+      } catch {
+        // If parsing fails, don't show balance error
+      }
+    }
+
+    return ''
+  })
+
+  // Check if trade button should be disabled
+  const isTradeDisabled = computed(
+    () =>
+      !supportedNetwork.value ||
+      !isMarketOpen.value ||
+      !isSelectedAssetTradeable.value ||
+      !(
+        fromAmount.value !== '' &&
+        fromAmount.value !== '0' &&
+        fromAmountError.value === '' &&
+        toAmount.value !== '0'
+      ) ||
+      isLoadingQuote.value,
+  )
+
+  return {
+    hasPreQuoteError,
+    fromAmountError,
+    isTradeDisabled,
+  }
+}
